benchmark/ML workflows/resnet/resnet/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

onnxruntime
numpy
Pillow
onnx
opencv-python-headless
line-profiler